requirements.txt
pyyaml==6.0.1
jinja2==3.1.4
markdown==3.6
